Add WowCurrency and WowCurrencyCategory models and populate them (#449)
- Start reworking CacheStaticJob

+++ apps/backend/Models/Data/DumpCurrencyCategory.cs
@@ -5,11 +5,11 @@ namespace Wowthing.Backend.Models.Data
 {
     public class DumpCurrencyCategory
     {
-        public int ID { get; set; }
+        public short ID { get; set; }
 
-        public int ExpansionID { get; set; }
-        public int Flags { get; set; }
-        
+        public short ExpansionID { get; set; }
+        public short Flags { get; set; }
+
         [Name("Name_lang")]
         public string Name { get; set; }
     }
